Edited bugs for SalaryLoand Schedule Entry

+++ Common/BusinessLayer/Department.vb
@@ -18,7 +18,7 @@ Namespace BusinessLayer
         Public Property IdNo() As Int16
         Public Property DepartmentCode() As String
         Public Property DepartmentName() As String
-        Public Property ParentIdNo() As Integer?
+        Public Property ParentIdNo() As Int16?
         Public Property DepartmentNameAra() As String
         Public Property Notes() As String
         Public Property RevCostCenterIdNo() As Int16
